init godesk

+++ view/src/com.jenkin.view/wallpaper/WallpaperChanger.go
@@ -10,27 +10,12 @@ import (
 	"os"
 	"path/filepath"
 	"syscall"
-	"time"
 	"unsafe"
 )
 
 const (
-	UserAgent      = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/86.0.4240.75 Safari/537.36"
-	BingHomeURL    = "https://cn.bing.com"
 	CurrentPathDir = "cache/"
 )
-
-const (
-	Size1k string = "1920,1080"
-	Size2k string = "2560,1440"
-	Size4k string = "3840,2160"
-)
-
-// ImageSize 图片大小
-type ImageSize struct {
-	w string
-	h string
-}
 
 func init() {
 	_ = os.Mkdir(CurrentPathDir, 0755)
@@ -55,33 +40,60 @@ func setWindowsWallpaper(imagePath string) error {
 	return nil
 }
 
+func setSleepWallpaper(imagePath string) error {
+
+	dll := syscall.NewLazyDLL("user32.dll")
+
+	proc := dll.NewProc("SystemParametersInfoW")
+	_t, _ := syscall.UTF16PtrFromString(imagePath)
+	ret, _, _ := proc.Call(20, 1, uintptr(unsafe.Pointer(_t)), 0x1|0x2)
+	if ret != 1 {
+		return errors.New("系统调用失败")
+	}
+	return nil
+}
+
+// 判断所给路径文件/文件夹是否存在
+func Exists(path string) bool {
+	_, err := os.Stat(path) //os.Stat获取文件信息
+	if err != nil {
+		if os.IsExist(err) {
+			return true
+		}
+		return false
+	}
+	return true
+}
+
 // DownloadImage 下载图片,保存并返回保存的文件名的绝对路径
 func DownloadImage(imageURL string) (string, error) {
-
-	client := http.Client{}
-
-	request, err := http.NewRequest("GET", imageURL, nil)
-	if err != nil {
-		return "", err
-	}
-
-	response, err := client.Do(request)
-	if err != nil {
-		return "", err
-	}
-	body, err := ioutil.ReadAll(response.Body)
-	if err != nil {
-		return "", err
-	}
-
-	day := time.Now().Format("2006-01-02")
-
 	fileName := EncodeMD5(imageURL)
-	path := CurrentPathDir + fmt.Sprintf("[%s]%s", day, fileName) + ".jpg"
+	path := CurrentPathDir + fmt.Sprintf("%s", fileName) + ".jpg"
+	fmt.Println("校验图片是否已存在", path)
+	exist := Exists(path)
+	if !exist {
 
-	err = ioutil.WriteFile(path, body, 0755)
-	if err != nil {
-		return "", err
+		client := http.Client{}
+
+		request, err := http.NewRequest("GET", imageURL, nil)
+		if err != nil {
+			return "", err
+		}
+
+		response, err := client.Do(request)
+		if err != nil {
+			return "", err
+		}
+		body, err := ioutil.ReadAll(response.Body)
+		if err != nil {
+			return "", err
+		}
+		err = ioutil.WriteFile(path, body, 0755)
+		if err != nil {
+			return "", err
+		}
+	} else {
+		fmt.Println("壁纸：", fileName, "已存在,不用下载")
 	}
 	absPath, err := filepath.Abs(path)
 	if err != nil {
